add __log_get_data and __log_save_data as a wrapper around internals of struct log to capture and save data

+++ main.c
@@ -16,7 +16,7 @@ int main(int argc, char * argv[])
         sm.sample_interval.sec   = 0 ;
 
 	tick = __calculate_tick(&sm);
-	printf("tick = %d\n",tick);
+	printf("tick = %ld\n",(unsigned long int)tick);
 
 	exit(EXIT_SUCCESS);
 }
